+++ lib/src/extensions/paragraph.dart
@@ -2,6 +2,6 @@ import 'package:powerd6/models.dart';
 
 extension ParagraphContent on Paragraph {
   String toHtml() {
-    return description.map((e) => '<p>$e</p>').join();
+    return lines.map((e) => '<p>$e</p>').join();
   }
 }